END DAY 5

+++ routes/coworkingRoutes.js
@@ -2,19 +2,12 @@ const express = require('express')
 const router = express.Router()
 // const { Op } = require('sequelize')
 const { Coworking } = require('../db/sequelizeSetup')
+const {findAllCoworkings} = require('../controllers/coworkingControllers')
+
 
 router
     .route('/')
-    .get((req, res) => {
-        console.log(req.query);
-        Coworking.findAll()
-            .then((results) => {
-                res.json(results)
-            })
-            .catch(error => {
-                res.json(error.message)
-            })
-    })
+    .get(findAllCoworkings)
     .post((req, res) => {
         const newCoworking = { ...req.body }
 
@@ -53,7 +46,14 @@ router
         })
             .then((result) => {
                 if (result > 0) {
-                    res.json({ message: 'Le coworking a bien été mis à jour.', data: result })
+                    Coworking.findByPk(req.params.id)
+                    .then ((coworking) =>{
+                        res.json({ message: 'Le coworking a bien été mis à jour.', data: coworking })
+                    })
+                    .catch ((error) =>{
+                        res.json({ mesage: `Une erreur est survenue`, data: error.message })
+                    })
+                   
                 } else {
                     res.json({ message: `Aucun coworking n'a été mis à jour.` })
                 }
@@ -64,13 +64,32 @@ router
     }) 
 
     .delete((req, res) => {
-        Coworking.destroy({ where: { id: req.params.id } })
-            .then((result) => {
-                res.json({ mesage: `Le coworking a bien été supprimé.`, data: result })
+        //A. Vérifie que l'ID passé en req.params.id renvoie bien une ligne de notre table.
+        Coworking.findByPk(req.params.id)
+            .then ((coworking)=>{
+                //B. Si un coworking correspond à l'ID alors on exécute la méthode .destroy()
+                if (coworking) {
+                    Coworking.destroy({ where: { id: req.params.id } })
+                        //C. Si le coworking est bien supprimé, on affiche un message avec comme data le coworking récupéré dans le .finByPk()
+                        .then((result) => {
+                            res.json({ mesage: `Le coworking a bien été supprimé.`, data: coworking })
+                        })
+                        // D. Si la suppression a échoué, on retourne une réponse à POSTMAN
+                        .catch((error) => {
+                            res.json({ mesage: `La suppression a échoué.`, data: error.message })
+                 })
+                } else {
+                    // B. Si aucun coworking ne correspond à l'ID alos on retourne une réponse à POSTMAN
+                    res.json({ mesage: `Aucun coworking trouvé`})
+                }
+
             })
+             // E. Si une erreur est survenue dès le findByPk, on retourne une réponse à POSTMAN
             .catch((error) => {
-                res.json({ mesage: `La suppression a échoué.`, data: error.message })
+                res.json({ mesage: `La requête n'a pas aboutie.` })
+
             })
+        
     })
